+++ src/main/java/com/mycompany/election/Candidat.java
@@ -2,7 +2,7 @@ package com.mycompany.election;
 
 public class Candidat extends Personne{
 
-	public Candidat(String n, String p, Float[] o) {
+	public Candidat(String n, String p, Double[] o) {
 		super(n, p, o);
 	}
