+++ src/main/java/com/mycompany/election/Electeur.java
@@ -2,7 +2,7 @@ package com.mycompany.election;
 
 public class Electeur extends Personne{
 
-	public Electeur(String n, String p, Float[] o) {
+	public Electeur(String n, String p, Double[] o) {
 		super(n, p, o);
 	}
